+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 - Added environment variables to the variables if `--no-cache` is not set
 - Added more things todo in `README.md`
 - Added the clean commands in `README.md`
+
+## Version 1.0.2
+- Added compiler caching
+- Added C(XX)/ASM flags
+- Split `LINKER` into `STATIC` and `SHARED`
+- Print error msgs in `sys.stderr`
+  
+